Extract common array prim code
